component 나누기 작업
